outsourced hardcoded functionality of UserInputManager into flexible functions to facilitate the following implementation of Tile Definition selection by keyboard inputs
REMARKS:
In "keyboard::only", the tile placement did not work at first: A single press of "Q" or "ENTER" did not place a tile, but a quick double press will. After the double press, also the single key press places a tile -> fixed by updating the current grid index to the last grid index within boundaries whenever the Input Method changes. The reason was that currently, the Input Method change is done via a button in the Debug GUI. If switching the Input Method by mouse, one has to exit the grid (which turns the current grid index invalid). This information is persistent even after the input method change and will prevent the tile to be placed due to the out of bounds condition.