integrated sliders partially into UserSettingsManager, adapted margins and fixed memory leak bugs
CHANGES:
- Found and fixed memory leak bugs
- Integrated sliders of new Settings Submenu into UserSettingsManager Logic

TO-DO:
- Set silders to correct default value
- Overhaul margins

REMARK:
The memory leak was twofold:
1) "Resources still in use at exit"
2) "There are still MemoryPool allocs in use at exit!"

2) was caused by not freeing an instance of TCEHSliderCluster.new() with queue_free() after usage. Since this is a class and not registered as a node in the Scene Tree, it is not garbage collected by Godot on its own. Manually applying queue_free() after being finished with it got rid of "There are still MemoryPool allocs in use at exit!" Error and reduced the "Unclaimed Strings Database Leak" from 249 to 131.

The remaining 131 instances of 1) was created by the UserInputManager and its class based contextual logic. Since it is class based, it is also not registered in the Scene Tree and therefor not automatically garbage collected. This could be fixed by adding

func _exit_tree() -> void:
	self._logic.queue_free()

to the UserInputManager script. Some additional safety checks were added to ensure that there will be no crash if no contextual logic is attached (e.g. standalone testing of menus).